+++ nflow-engine/src/test/java/io/nflow/engine/workflow/definition/WorkflowSettingsTest.java
@@ -10,12 +10,16 @@ import static org.joda.time.DateTimeUtils.currentTimeMillis;
 import static org.joda.time.DateTimeUtils.setCurrentMillisFixed;
 import static org.joda.time.DateTimeUtils.setCurrentMillisSystem;
 
+import java.util.function.BiFunction;
 import java.util.function.BooleanSupplier;
 
 import org.joda.time.DateTime;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.slf4j.event.Level;
+
+import io.nflow.engine.exception.StateProcessExceptionHandling;
 
 public class WorkflowSettingsTest {
   DateTime now = new DateTime(2014, 10, 22, 20, 44, 0);
@@ -46,12 +50,13 @@ public class WorkflowSettingsTest {
   public void errorTransitionDelayIsBetweenMinAndMaxDelay() {
     int maxDelay = 1_000_000;
     int minDelay = 1000;
-    WorkflowSettings s = new WorkflowSettings.Builder().setMinErrorTransitionDelay(minDelay).setMaxErrorTransitionDelay(maxDelay).build();
+    WorkflowSettings s = new WorkflowSettings.Builder().setMinErrorTransitionDelay(minDelay).setMaxErrorTransitionDelay(maxDelay)
+        .build();
     long prevDelay = 0;
-    for(int retryCount = 0 ; retryCount < 100 ; retryCount++) {
-      long delay  = s.getErrorTransitionActivation(retryCount).getMillis() - now.getMillis();
-      assertThat(delay, greaterThanOrEqualTo((long)minDelay));
-      assertThat(delay, lessThanOrEqualTo((long)maxDelay));
+    for (int retryCount = 0; retryCount < 100; retryCount++) {
+      long delay = s.getErrorTransitionActivation(retryCount).getMillis() - now.getMillis();
+      assertThat(delay, greaterThanOrEqualTo((long) minDelay));
+      assertThat(delay, lessThanOrEqualTo((long) maxDelay));
       assertThat(delay, greaterThanOrEqualTo(prevDelay));
       prevDelay = delay;
     }
@@ -101,4 +106,48 @@ public class WorkflowSettingsTest {
     assertThat(supplier.getAsBoolean(), is(false));
   }
 
+  @Test
+  public void defaultExceptionAnalyzer() {
+    WorkflowSettings s = new WorkflowSettings.Builder().build();
+
+    StateProcessExceptionHandling exceptionHandling = s.analyzeExeption(TestWorkflow.State.begin, new Throwable());
+    assertThat(exceptionHandling.isRetryable, is(true));
+    assertThat(exceptionHandling.logLevel, is(Level.ERROR));
+    assertThat(exceptionHandling.logStackTrace, is(true));
+
+    exceptionHandling = s.analyzeExeption(TestWorkflow.State.begin, new NonRetryableException());
+    assertThat(exceptionHandling.isRetryable, is(false));
+    assertThat(exceptionHandling.logLevel, is(Level.ERROR));
+    assertThat(exceptionHandling.logStackTrace, is(true));
+  }
+
+  @Test
+  public void customExceptionAnalyzer() {
+    WorkflowSettings s = new WorkflowSettings.Builder()
+        .setExceptionAnalyzer((state, thrown) -> new StateProcessExceptionHandling.Builder()
+        .setLogLevel(Level.INFO).setRetryable(true).setLogStackTrace(false).build()).build();
+
+    StateProcessExceptionHandling exceptionHandling = s.analyzeExeption(TestWorkflow.State.begin, new NonRetryableException());
+    assertThat(exceptionHandling.isRetryable, is(true));
+    assertThat(exceptionHandling.logLevel, is(Level.INFO));
+    assertThat(exceptionHandling.logStackTrace, is(false));
+  }
+
+  @Test
+  public void defaultExceptionAnalyzerIsUsedWhenCustomAnalyzerFails() {
+    BiFunction<WorkflowState, Throwable, StateProcessExceptionHandling> failingExceptionAnalyzer = (state, thrown) -> {
+      throw new IllegalStateException("fail");
+    };
+    WorkflowSettings s = new WorkflowSettings.Builder().setExceptionAnalyzer(failingExceptionAnalyzer).build();
+
+    StateProcessExceptionHandling exceptionHandling = s.analyzeExeption(TestWorkflow.State.begin, new NonRetryableException());
+    assertThat(exceptionHandling.isRetryable, is(false));
+    assertThat(exceptionHandling.logLevel, is(Level.ERROR));
+    assertThat(exceptionHandling.logStackTrace, is(true));
+  }
+
+  @NonRetryable
+  static class NonRetryableException extends RuntimeException {
+    private static final long serialVersionUID = 1L;
+  }
 }